+++ demo/test/feature_test.py
@@ -4,11 +4,19 @@ import rubato as rb
 rb.init(
     res=(1080, 1080),
     window_size=(2000, 2000),
-    fullscreen=True,
-    maximize=True,
+    # fullscreen=True,
+    # maximize=True,
 )
 
 main = rb.Scene()
-main.add(rb.wrap(rb.Rectangle(200, 200)))
+a = rb.Image("art/avocado.png", offset=(-250, 0))
+
+a.set_alpha(50)
+a.set_colorkey(rb.Color.from_hex("#ecf986"))
+
+b = a.clone()
+b.offset.x += 500
+
+main.add(rb.wrap([a, b]))
 
 rb.begin()